grcqt: Finished combining the separate views/controllers
