+++ grc/gui_qt/components/__init__.py
@@ -1,1 +1,8 @@
+
 from .block_library import BlockLibrary
+from .console import Console
+from .flowgraph import Flowgraph
+from .project_manager import ProjectManager
+
+# Import last since there are dependencies
+from .window import MainWindow
