+++ grc/gui_qt/components/flowgraph.py
@@ -1,29 +1,53 @@
+# Copyright 2014-2020 Free Software Foundation, Inc.
+# This file is part of GNU Radio
+#
+# GNU Radio Companion is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# GNU Radio Companion is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA
+
+from __future__ import absolute_import, print_function
+
+# Standard modules
+import logging
+
+import xml.etree.ElementTree as ET
+
+from ast import literal_eval
+
+# Third-party modules
 import six
 
-# standard modules
-import xml.etree.ElementTree as ET
-from   ast import literal_eval
-
-# third-party modules
 from PyQt5 import QtGui, QtCore, QtWidgets
 from PyQt5.QtCore import Qt
 
-# custom modules
-from . block import Block
-
+# Custom modules
+from .canvas.block import Block
 from .. import base
 
+# Logging
+log = logging.getLogger(__name__)
 
 DEFAULT_MAX_X = 1280
 DEFAULT_MAX_Y = 1024
 
+
 # TODO: Combine the scene and view? Maybe the scene should be the controller?
-class FlowgraphScene(QtWidgets.QGraphicsScene, base.Controller):
+class FlowgraphScene(QtWidgets.QGraphicsScene, base.Component):
     def __init__(self, *args, **kwargs):
         super(FlowgraphScene, self).__init__()
         self.isPanning    = False
         self.mousePressed = False
-        
+
     def dragEnterEvent(self, event):
         if event.mimeData().hasUrls:
             event.setDropAction(Qt.CopyAction)
@@ -76,10 +100,10 @@ class FlowgraphScene(QtWidgets.QGraphicsScene, base.Controller):
                         key = p['label']
                         value = p.get('default', '') # just show default value for now
                         params.append((key, value))
-                    
+
                 # Tell the block where to show up on the canvas
                 attrib = {'_coordinate':(cursor_pos.x(), cursor_pos.y())}
-                
+
                 new_block = Block(block_key, block.label, attrib, params)
                 self.addItem(new_block)
 
@@ -89,7 +113,7 @@ class FlowgraphScene(QtWidgets.QGraphicsScene, base.Controller):
                 return QtGui.QStandardItemModel.dropMimeData(self, data, action, row, column, parent)
         else:
             event.ignore()
-    
+
     def mousePressEvent(self,  event):
         if event.button() == Qt.LeftButton:
             self.mousePressed = True
@@ -131,11 +155,28 @@ class FlowgraphScene(QtWidgets.QGraphicsScene, base.Controller):
         print("You double clicked on a block")
 
 
-        
-        
-        
 
-class Flowgraph(QtWidgets.QGraphicsView, base.Controller): # added base.Controller so it can see platform
+    def createActions(self, actions):
+        log.debug("Creating actions")
+
+        '''
+        # File Actions
+        actions['save'] = Action(Icons("document-save"), _("save"), self,
+                                shortcut=Keys.New, statusTip=_("save-tooltip"))
+
+        actions['clear'] = Action(Icons("document-close"), _("clear"), self,
+                                         shortcut=Keys.Open, statusTip=_("clear-tooltip"))
+        '''
+
+    def createMenus(self, actions, menus):
+        log.debug("Creating menus")
+
+    def createToolbars(self, actions, toolbars):
+        log.debug("Creating toolbars")
+
+
+
+class Flowgraph(QtWidgets.QGraphicsView, base.Component): # added base.Component so it can see platform
     def __init__(self, parent, filename=None):
         super(Flowgraph, self).__init__()
         self.setParent(parent)
@@ -151,9 +192,46 @@ class Flowgraph(QtWidgets.QGraphicsView, base.Controller): # added base.Controll
 
         self.setScene(self.scene)
         self.setBackgroundBrush(QtGui.QBrush(Qt.white))
-        
+
         self.isPanning    = False
         self.mousePressed = False
+
+        '''
+        QGraphicsView.__init__(self, flow_graph, parent)
+        self._flow_graph = flow_graph
+
+        self.setFrameShape(QFrame.NoFrame)
+        self.setRenderHints(QPainter.Antialiasing |
+                            QPainter.SmoothPixmapTransform)
+        self.setAcceptDrops(True)
+        self.setAlignment(Qt.AlignLeft | Qt.AlignTop)
+        self.setSceneRect(0, 0, self.width(), self.height())
+
+        self._dragged_block = None
+
+        #ToDo: Better put this in Block()
+        #self.setContextMenuPolicy(Qt.ActionsContextMenu)
+        #self.addActions(parent.main_window.menuEdit.actions())
+        '''
+
+
+    def createActions(self, actions):
+        log.debug("Creating actions")
+
+        '''
+        # File Actions
+        actions['save'] = Action(Icons("document-save"), _("save"), self,
+                                shortcut=Keys.New, statusTip=_("save-tooltip"))
+
+        actions['clear'] = Action(Icons("document-close"), _("clear"), self,
+                                         shortcut=Keys.Open, statusTip=_("clear-tooltip"))
+        '''
+
+    def createMenus(self, actions, menus):
+        log.debug("Creating menus")
+
+    def createToolbars(self, actions, toolbars):
+        log.debug("Creating toolbars")
 
     def readFile(self, filename):
         tree = ET.parse(filename)
@@ -174,10 +252,13 @@ class Flowgraph(QtWidgets.QGraphicsView, base.Controller): # added base.Controll
                     params.append((key, value))
 
             # Find block in tree so that we can pull out label
-            block = self.platform.blocks[block_key]
-                
-            new_block = Block(block_key, block.label, attrib, params)
-            self.scene.addItem(new_block)
+            try:
+                block = self.platform.blocks[block_key]
+
+                new_block = Block(block_key, block.label, attrib, params)
+                self.scene.addItem(new_block)
+            except:
+                log.warning("Block '{}' was not found".format(block_key))
 
         # This part no longer works now that we are using a Scene with GraphicsItems, but I'm sure there's still some way to do it
         #bounds = self.scene.itemsBoundingRect()
@@ -244,9 +325,42 @@ class Flowgraph(QtWidgets.QGraphicsView, base.Controller): # added base.Controll
                 self.setCursor(Qt.ArrowCursor)
         else:
             super(Flowgraph, self).keyPressEvent(event)
-            
+
     def mouseDoubleClickEvent(self, event):
         # This will pass the double click event to the scene
         super(Flowgraph, self).mouseDoubleClickEvent(event)
 
 
+    '''
+    def wheelEvent(self, event):
+        # TODO: Support multi touch drag and drop for scrolling through the view
+        if event.modifiers() == Qt.ControlModifier:
+            factor = 1.2
+            if event.delta() < 0 :
+                factor = 1.0 / factor
+            self.scale(factor, factor)
+        else:
+            QGraphicsView.wheelEvent(self, event)
+
+    def dragEnterEvent(self, event):
+        key = event.mimeData().text()
+        self._dragged_block = self._flow_graph.add_new_block(
+            str(key), self.mapToScene(event.pos()))
+        event.accept()
+
+    def dragMoveEvent(self, event):
+        if self._dragged_block:
+            self._dragged_block.setPos(self.mapToScene(event.pos()))
+            event.accept()
+        else:
+            event.ignore()
+
+    def dragLeaveEvent(self, event):
+        if self._dragged_block:
+            self._flow_graph.remove_element(self._dragged_block)
+            self._flow_graph.removeItem(self._dragged_block)
+
+    def dropEvent(self, event):
+        self._dragged_block = None
+        event.accept()
+    '''
