--- grc/gui_qt/controllers/__init__.py
@@ -1,4 +1,0 @@
-from . window import MainWindow
-from . reports import Reports
-from . drawing import DrawingController
-from . projects import ProjectsController
