--- grc/gui_qt/controllers/drawing.py
@@ -1,5 +1,0 @@
-class DrawingController(object):
-    def __init__(self, view):
-
-        # Load the main view class and initialize QMainWindow
-        self._view = view
