--- grc/gui_qt/controllers/helpers/__init__.py
@@ -1,1 +1,0 @@
-from . import Qt
