--- grc/gui_qt/views/__init__.py
@@ -1,4 +1,0 @@
-from . window import MainWindow
-from . reports import Reports
-from . projects import ProjectsView
-from . flowgraph import Flowgraph
